tests/operator: split upgrade tests into upgrade.go
Move the upgrade DescribeTable and all upgrade-specific helpers
(getUpstreamReleaseAssetURL, detectLatestUpstreamOfficialTag, getTagHint,
installOperator, expectVirtOperatorPodsToTerminate, and the
generatePrevious* functions) into upgrade.go.

The BeforeEach in upgrade.go initialises per-spec state; a matching
AfterEach restores the KubeVirt installation to a stable state after
each (potentially destructive) test, mirroring the structure of the
original operator.go.
